refactor random word function to not require msg
